requirements.txt
Django==1.7.2
django-allauth==0.19.0
google-api-python-client==1.3.1
httplib2==0.9
mysqlclient==1.3.4
oauth2client==1.4.5
oauthlib==0.7.2
Pillow==2.7.0
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-openid==2.2.5
requests==2.5.1
requests-oauthlib==0.4.2
rsa==3.1.4
simplejson==3.6.5
six==1.9.0
Unipath==1.0
uritemplate==0.6
